Adaption Sight seeking angle

+++ src/main/java/org/xflash/utils/AngleUtils.java
@@ -4,7 +4,7 @@ package org.xflash.utils;
  */
 public class AngleUtils {
 
-    public static double getTargetAngle(int x, int y, float centerX, float centerY) {
+    public static double getTargetAngle(double x, double y, double centerX, double centerY) {
         double theta = Math.atan2(y - centerY, x - centerX);
         return theta < 0 ? theta + Math.PI * 2 : theta;
     }
